Don't kill the grpc server process when the parent process ends
Summary: Still writing a test that specifically captures the behavior change here, but putting out for review while I do that to ensure that we're on the same page about the plan. The idea is that we no longer kill the server process when the dagster host process ends, but we do send a shutdown signal when the ephemeral client shuts down. We'll still need a heartbeat to ensure that the server process doesn't stick around forever if the dagster process suddenly crashes, which max is working on separately.

Test Plan: New test case coming

Reviewers: max, alangenfeld, sashank

Differential Revision: https://dagster.phacility.com/D4096