Throw from sync_get_external_execution_plan rather than relying on callsites to check error data
Summary:
I added ExecutionPlanSnapshotErrorData in https://dagster.phacility.com/D4031 so that we could get clear data from grpc/the cli api about errors across process boundaries. The problem with that is that I made is so that it's returned from get_external_execution_plan, so every callsite needs to check that error code. Instead, make that method throw, and callsites that want can catch exceptions and pull out the error info from i
t.

Test Plan:
Load a pipeline in dagit with missing config that requires it, error message is clear again

Reviewers: max, schrockn, alangenfeld, sandyryza