Fix invalid workspace example yaml in docs
Summary:
The schema doesn't allow you to put location_name adjacent to python_package like this, it needs to be a child. Whether it should be possible to do it this way is another question (I'd argue it should go alongside executable_path and target), but in the meantime, make the YAML compile.

Test Plan:
Paste example YAML into a workspace.yaml file, load it in dagit, no more schema error

Reviewers: sandyryza, yuhan, alangenfeld