made it unnecessary to require run config for intermediate storage when the intermediate_storage has no config schema.
Summary: Previously, a run config was always necessary to provide an intermediate storage option to a pipeline. This change makes it so that if the intermediate storage definition for a given pipeline has no config schema, it will not require a run config.

Test Plan: Added a test to ensure correct behavior when no run config is provided.

Reviewers: sandyryza

Differential Revision: https://dagster.phacility.com/D4178

explicitly defined failure modes

made code reuse easier

made in memory the only default intermediate storage. Made detection slightly less smart so to be easier for users.

removed changes of default intermediate storage options to preserve backward compatibility

got rid of unused utility fxn

when only one system storage is defined and it requires no config schema, it can be used with no run config provided.

linted

got rid of default value in config for storages. With the new changes, it would never be used.

re-introduced default parameter for IntermediateStorageConfig, to handle the case where no run config is passed for either storage or intermediatestorage.

refactored tests to be more succinct

docstring formatting

(Depends on D4178) Enabled errors with storage config to populate in dagit. Added default values to storage definitions when it is not erroneous.

Summary: Previously, a run config was always necessary to provide an intermediate storage option to a pipeline. This change makes it so that if the intermediate storage definition for a given pipeline has no config schema, it will not require a run config.

Test Plan: Added a test to ensure correct behavior when no run config is provided.

Reviewers: sandyryza

Differential Revision: https://dagster.phacility.com/D4215

deferred intermediate storage to use default system storage. This is a temporary fix to preserve backwards compatibility with the system storage API.

further changes to ensure backwards compatibility

added tests to ensure pipeline definitions complete and/or error out correctly

fixed docstrings

got rid of rebase artifact.

reverted objects.py since its changes were no longer necessary. Changed tests accordingly.

linted