made it unnecessary to require run config for intermediate storage when the intermediate_storage has no config schema.
Summary: Previously, a run config was always necessary to provide an intermediate storage option to a pipeline. This change makes it so that if the intermediate storage definition for a given pipeline has no config schema, it will not require a run config.

Test Plan: Added a test to ensure correct behavior when no run config is provided.

Reviewers: sandyryza

Subscribers: