(Depends on D4178) Enabled errors with storage config to populate in dagit. Added default values to storage definitions when it is not erroneous.
Summary: Previously, a run config was always necessary to provide an intermediate storage option to a pipeline. This change makes it so that if the intermediate storage definition for a given pipeline has no config schema, it will not require a run config.

Test Plan: Added a test to ensure correct behavior when no run config is provided.

Reviewers: sandyryza

Differential Revision: https://dagster.phacility.com/D4215

deferred intermediate storage to use default system storage. This is a temporary fix to preserve backwards compatibility with the system storage API.

further changes to ensure backwards compatibility

added tests to ensure pipeline definitions complete and/or error out correctly

fixed docstrings

got rid of rebase artifact.