Use instance_for_test in more places (especially in the run launcher tests)
Summary:
This is kind of a shot in the dark to deal with the pytest hanging issues we've been having during teardown. Use an instance per test instead of a module fixture, and clean up all runs between tests using the . If nothing else hopefully this will help us isolate it to a specific test.

Reviewers: max, alangenfeld

Test Plan: BK, Azure