[partition-backfill-6] ui for partitions backfill
Summary: This diff adds a pipeline-scoped partitions dashboard, it contains the partition-based
longitudinal view, as well as a new partition backfill launch section.

This enables the user to launch full partition executions (equivalent to the CLI), as well as
backfills relying on the intermediate store of previous runs (either successful, or from failure).
The latter is based off of the re-execution options from an existing run.

Test Plan: bk

Reviewers: alangenfeld, bengotow, sashank

Subscribers: