hook-prototype post-hoc hook accumulation but limited composite support
Summary:
a hook can be attached to any of the following objects
  * PipelineDefinition
  * Solid (solid invocation)
and we gather all the hooks at hook invocation time so we don't need to copy the pipeline_def.dependencies to make the hooks applied to each solid invocation inside a pipeline_def

this would support hooks on pipelines with composites (D4220 will error out)

however, we still can't do
```
@pipeline
def foo_pipeline():
    a_composite.with_hooks({slack_on_failure})()
```
because to attach a set of hooks to composite, we will likely treat the `CompositeSolidDefinition` simliarly to `PipelineDefinition`, meaning we will add the hook_defs to a copy of the composite solid def. But, `CompositeSolidDefinition` also behaves like a `SolidDefinition` whose name has to be unique -- we will have to make a copy with different names.
- this is a similar case to `@configured` (where we create a copy of the solid definition to pass through the new configuration) but unfortunately also different: semantically, `@configured` makes more sense to be attached to the definition level while it makes more sense to apply hooks on instance level. So configured can create a copy with the same name as we don't expect multiple copies to be created, but the `composite_solid_a.with_hooks()()` could conflict with `composite_solid_a()` as one of a copy of the other :/

Test Plan:
bk
Reviewers:

Subscribers: