Add docs and release notes explaining how to use the managed gRPC feature
Summary:
This is approaching the state where we can start to gather feedback on it I think? Let me know if you think it's still too early.

I couched it in terms of execution time but I should probably run it on some representative pipelines first to verify that there is in fact a speedup.

Reviewers: schrockn, alangenfeld, max, sashank

Test Plan:
View docs and CHANGES.md page