Don't make scaffold_config command ue externalpipeline
Summary:
Much like the execute command before D4156, this command is in a weird state where it loads things out-of-process but then requires it to be a host process. There isn't enough info on ExternalPipeline currently to fully re-create it, and even if there was it would be a bunch of work to port it for a relatively obscure (?) command. So this makes it require a host process, much like the pipelin execute command.

Reviewers: alangenfeld, schrockn

Test Plan: BK