Don't return the run from RunLauncher.launch_run
Summary:
Right now all of our run launchers don't modify the run that's passed in and just return it bakc. Returning the run from this method creates the inaccurate impression that it has been updated in some way to reflect the launch of the run, but most changes that happen to the run will be asynchronous (e.g. updating the run status), so I think we should just not return anything from this method. The caller has the run ID and can fetch it again from the run storage if they want an updated copy.

That said this is a minor breaking change and is not urgent so we could also hold off.

Test Plan: BK

Reviewers: alangenfeld

Subscribers: