[Fix for Issue #1563] added warning when solid not invoked from composition function
Summary: Added a warning that fires when a solid is defined, but not invoked, from a composition fxn such as @pipeline.

Test Plan: Tested using example provided in issue.

Differential Revision: https://dagster.phacility.com/D4322

added check for verifying actually in composition fxn

added a test for solids not invoked

changed where warning was checked.

changed warning msg check