Delay interrupts that arrive while opening the compute log subprocesses
Summary:
If an interrupt comes in as these processes are being opened, they stay open, but we have no way to terminate them. Use delay_interrupts() to wait until we have such a reference for any surprise exceptions to come in.

Additionally, fix a place where we were calling multiprocess.Process.kill(), which is a method that does not exist on all python versions.

Test Plan: BK several times, no more test failures in the interrupt tests

Reviewers: alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D4358