Join GRPC server threads on termination
Summary: Right now we are not joining on any of the threads we create in the grpc server process. Add an explicit join() step for each of them. My suspicion is this may be contributing to the lingering segfaults on some grpc test runs, we were able to reduce them dramatically by doing something similar on the grpc client heartbeat thread.

Test Plan: BK

Reviewers: alangenfeld, max

Differential Revision: https://dagster.phacility.com/D4364