RFC: Check that pipeline is valid
Summary: Because we're good at stuff we were actually never checking that the pipeline name was valid.

This is potentially a breaking change for users but is the right thing
long term. I actually think that we should also make pipeline name
required while we're at it and get rid of the <<unnamed>> BS. All the
unnamed stuff predates the decorator syntax and is less necessary now.

Test Plan: BK

Reviewers: prha, alangenfeld

Differential Revision: https://dagster.phacility.com/D4373