Make DagsterInstance a contextmanager for resource cleanup
Summary:
Rather than relying on __del__ or asking all DagsterInstance callsites to remember to call .dispose(), use a contextmanager and make the default mode of operation when creating a DagsterInstance to wrap it in a with. This should help with sporadic classes in postgres tests in teardown when trying to kill the watcher thread.

Assuming this seems reaonable, A separate diff will actually migrate the major DagsterInstance.get() callsites to use this - one is provided here as an example.

Test Plan: BK

Reviewers: schrockn, alangenfeld, catherinewu

Differential Revision: https://dagster.phacility.com/D4374