Remove stray DagsterInstance.ephemeral() call in dagster/core
Summary: I think this is the only place left in the main dagster module (other than tests) where we're constructing a raw DagsterInstance vs. passing it in or reconstructing it from an instance ref. I'm not actually sure what exactly we're using the DagsterInstance for (I think it might just be for loggers?), but as long as we're using it it's better to pass it in explicitly (even if the callsite constructs an ephemeral instance for it)

Test Plan: BK

Reviewers: sandyryza, alangenfeld, schrockn

Differential Revision: https://dagster.phacility.com/D4376