Show error toast on termination PythonError
Summary:

When a Python error is raised when terminating a run, the "Terminate" button doesn't appear to do anything and no error is surfaced.

Resolve this by ensuring that the `PythonError` is also handled in the mutation.

Test Plan:

Run an intentionall slow pipeline, with a `raise Exception('oh no')` in the termination codepath. Click "Terminate", verify that the error toast appears.

Reviewers:

catherinewu, bengotow, prha

Subscribers: