Don't show snapshot banner for current pipeline
Summary:

The historical snapshot banner appears on the pipeline display when loading a specific run, even when the relevant pipeline is current.

Check the ID in the URL path against the ID of the pipeline, and if they match, don't show the banner.

Test Plan:

Run a pipeline. Change the pipeline solids, run it again. View each run, verify that the stale pipeline run shows the banner but that the current one does not.

Reviewers:

alangenfeld, prha, bengotow

Subscribers: