Better handling for the case where gRPC server can't load the run being launched
Summary:
Makes a useful error be surfaced to the user when this happens, instead of returning a successful start and then immediately failing.

Also add a doc line to make it clear that we need to be able to load the run from storage.

Test Plan: New BK coverage, simulate failure by setting pipeline_run to Null and then launching in dagit

Reviewers: alangenfeld, max

Differential Revision: https://dagster.phacility.com/D4469