Make lazy-loading user code in the gRPC server a flag that's set in managed gRPC servers only
Summary:
In https://dagster.phacility.com/D4087 we made it so that the gRPC server waits until the first LoadRepositories call before it surfaces any load errors. This is useful for the managed gRPC case for surfacing errors clearly, but in the deployed gRPC case, we should probably surface them as soon as they start up the server. So make it a flag that is set in the managed/process-bound gRPC case, but not in the deployed case.

Separately, we should ensure that the stack trace that comes out of the managed gRPC list repository call is super actionable in dagit.

Resolves #2917

Test Plan: New and existing BK coverage

Reviewers: alangenfeld, sashank, sandyryza