Disable tabs for pipeline snapshots
Summary:

On a historical snapshot view for a pipeline, the navigation tabs don't currently make sense: they navigate you away from the snapshot to the current pipeline, which could cause confusion.

Instead, disable the tabs in the snapshot case, and provide a link to the current pipeline within the snapshot notice. Hopefully this helps clarify that there are no operations to perform on the snapshot.

Test Plan:

View snapshot, verify details above.

Reviewers:

alangenfeld, bengotow, prha

Subscribers: