added better documentation for step version resolution
Added extensive documentation to all version resolution computations.

Summary: The version resolution code was becoming super hard to understand. I wanted to nip this in the bud by codfying exactly how we are computing each version, so that what the code is doing makes more sense to an outside observer.

Test Plan: na

Reviewers: sandyryza, yuhan

Differential Revision: https://dagster.phacility.com/D4497