changed solid version to index by string-ified version of solid handle rather than name of solid - handles composite case
Summary: within an instance of EnvironmentConfig, solid config is indexed by the string of the solid handle. Most of the time, this boils down to just the name of the solid, but if the solid is either aliased or composite, then it will be different. Changed what resolve_step_versions indexed on to compensate for this.

Test Plan: Added a new test `test_step_versions_composite_solid` which tests for this case. Modeled after `test_good_override` in `test_config_mappings.py`, which was one of the failing tests that brough this to our attention.

Reviewers: sandyryza, yuhan

Differential Revision: https://dagster.phacility.com/D4498

called string method of step handle