Add tests for scheduler crash recovery
Summary:
Ensure that even if the scheduler crashes or is interrupted mid-schedule, we recover and create exactly one tick and exactly one run if the tick was successful.

The only way I could figure out to do this was a little invasive, by passing in a map to the scheduler function that can tell it to crash itself at various places. Open to alternatives for accomplishing that goal.

Also noted one place inline where there's still a possible race condition that launches the same run twice if we get super unlucky, probably worth considering a mitigation strategy for that (but it's tricky unless there's some way of making the launch operation transactional in some way). Maybe the gRPC server could do some tracking of this as well, and reject a run that has previously been launched on that server?

Test Plan: BK

Reviewers: alangenfeld, sashank

Differential Revision: https://dagster.phacility.com/D4517