[dagit] Use interactive Tags for log filter
Summary:
On a Run page logs section, the filter buttons are a little hard to identify as "on" or "off", as both states use gray.

To make the distinction clearer, and to more closely reflect the filter UI seen in Chrome devtools (which always seem pretty clear to me), use the blueprint `Tag` component with "primary" intent blue to indicate "on", and plain gray to indicate "off".

Test Plan: View log filter on a Run page, verify proper rendering of filter tags.

Reviewers: bengotow, alangenfeld, prha

Differential Revision: https://dagster.phacility.com/D4518