Properly clean up run launcher thread when instance shuts down
Summary:
Noticed this when testing the scheduler command - after each run there's often an exception in the watcher thread. .join() cleans up the thread, but we only run that in tests. Use the relatively new DagsterInstance contextmanager functionality to do a cleanup of the thread every time the instance shuts down (without waiting for the runs to complete like we do in join()).

Test Plan: Existing BK coverage, run scheduler command and see no more failures from hanging cleanup threads after each run. The failure that was happening before one is quite tricky to test in a unit test unfortunately because it only happened during interpreter teardown and occurred as a non-blocking error spew in another thread.