RFC [dagit] Highlight matches during logs text search
Summary:
A proposed change to the behavior of text-search filtering on the logs view for a Run, based on what seemed like some confusing UX during the debugging session on Zoom today.

Instead of hiding logs that don't match the text search, keep everything displayed, but *highlight* the matches. This way, you can see the matches in the context of the surrounding logs, rather than having the undo the search to see everything else that happened.

The highlighting is pretty subtle, so if this seems like a good direction, I'm certainly open to suggestions.

Test Plan: View Run page, perform text searches on logs. Verify that the filtering behaves correctly, and that logs render with the highlight.

Reviewers: bengotow, catherinewu, alangenfeld, prha, sashank

Differential Revision: https://dagster.phacility.com/D4522