Write non-error execution logs to stdout instead of stderr, only surface stderr from child processes in host processes like the scheduler
Summary:
The high-level goal here is to make the scheduler process output more legible - right now it surfaces every log from every execution it triggers, which is not very useful if you're just trying to monitor what's going on.

This happens because the scheduler opens user processes to execute scheduled runs, and they are very verbose.

Where I landed for now (Which I'm not positive is right) is to still include anything that was logged to stderr in the scheduler runs, and to change our default logger so that it writes INFO and DEBUG to stdout (It always confused me why we were writing all log output to stderr, although that appears to be the default Python logging behavior). Then only surface STDERR in host processes by default (the gRPC server already does this, but that's because it reads stdout to determine when the server started)

This would also affect console output from Dagit.

Does this seem reasonable? If so I will write some tests.

Test Plan: Pending based on initial feedback

Reviewers: alangenfeld, sandyryza, schrockn, sashank

Differential Revision: https://dagster.phacility.com/D4538