[dagit] Allow z precedence for split panels
Summary:

In cases where flyouts are rendered from split-pane contexts, we can end up with the flyouts rendering beneath the other panel in an undesirable way.

Allow setting which panel has "top" precedence as a quick fix for this. This is not a great (or even good) long-term solution, but it repairs the immediate issue for split panels.

Test Plan:

View solid selector modal on Playground tab for a pipeline. Verify that it visually sits above the bottom panel.

Reviewers:

bengotow, alangenfeld, sashank, prha

Subscribers: