[active execution] throw on lost steps
Summary: Currently we can end a pipeline in success despite losing communication with step processes since step failure events were not observed and no exception was thrown.

This diff takes the approach of throwing a new exception type in the event that the execution is complete but there were lost steps.

depends on D4591

Test Plan:
Need to think on how to get this in a pytest.
Manually commented out the step failure event in the multiprocess executor and used `error_monster` with a process crash to simulate lost step comms.

Reviewers: dgibson, catherinewu,

Subscribers: