[dagit] Allow reversing partition order on Playground
Summary:

Since the partitions dropdown list is pretty long, allow reversing the display so that you don't have to scroll all the way to the bottom or remember exactly what to search for in order to get to that part of the list.

Test Plan:

Video attached.

Reviewers:

bengotow, catherinewu, alangenfeld, sashank

Subscribers: