Change the default system for loading and executing user code to be gRPC instead of the CLI API
Summary:
Kind of an RFC. Alex and I (and hopefully others?) have been using this flag locally to deprecate the CLI API and do everything over gRPC for some time and have fixed a number of issues that came up, it seems pretty stable at this point. Is it time to flip the switch and make it the default? We still have an opt-out switch if new surprising issues come up.

Reviewers: alangenfeld, schrockn, max, nate

Test Plan: BK, launch runs in dagit and verify they are happening over gRPC