Enabled versioning for commonly used builtin types when loading inputs from config.
Summary: In order to enable versioning on externally-loaded inputs to a solid, it is necessary to implement a function that in some way converts the externally-loaded value into a version. This implements a basic function to do just that for the common builtin types (Bool, Int, Float, String).

Test Plan: Expanded testing infrastructure within `test_versioned_execution_plan.py` to run a test for externally loaded inputs against each of the builtin types that this fix works for, and incorporates the test for custom externally loaded types as well.

Reviewers: sandyryza, yuhan

Subscribers: