[dagit] Allow copying filter state to URL on Run page
Summary:
Resolves #3029

A handful of changes to the behavior of query parameters on the Run page in dagit:

- Allow setting filter values via URL params:
  - `steps` to set the step selection on the Gantt chart
  - `logs` to set log filter tokens
  - `levels` to set the levels filter on the logs view
- Add a "Copy URL" button to the logs toolbar to allow deep linking to the current filter state
- Add the ability to use a `query:` token in the logs filter, since it's used implicitly by step selection

I also made the jest config not insist on tests being in `__tests__` directories, since I think this is the kind of thing that adds friction to creating new tests (which we need more of).

Test Plan:
View a Run page. Modify filters, and click "Copy URL" in each case, then open that copied URL to a new tab and verify correct filter initialization.

- Change step selection, using wildcard characters and multiple steps. Verify proper filtering behavior.
- Change log filter tokens, verify same.
- Change levels, verify same.
- Change a combination of these, verify same.

Reviewers: bengotow, yuhan, catherinewu, sashank

Differential Revision: https://dagster.phacility.com/D4765