[dagit] Allow snapshots to re-run even if the snapshot ID is mismatched
Summary:
When viewing a run, we currently disable the re-execution button if the snapshot ID doesn't match the current snapshot ID for that pipeline in the active repo.

An issue here is that a run on a solid subselection will produce a different snapshot ID from the parent pipeline, which means this button will always end up being disabled when trying to re-run that subselection.

Instead, show a tooltip warning but allow the re-execution to occur.

Test Plan: View a historical run, verify that the warning appears and that I am able to re-execute the run.

Reviewers: bengotow, alangenfeld, sandyryza, max

Differential Revision: https://dagster.phacility.com/D4784