[dagit] Use parent snapshot ID to determine current-ness of snapshot
Summary:
In cases where a snapshot is a child of a parent snapshot, use that information to determine whether to tag the snapshot as "current". That is, if a parent snapshot is the current state of the pipeline, it should be labeled as "current".

- Add nullable `parentSnapshotId` to `PipelineSnapshot` in GraphQL
- Break apart PipelineNav
  - Query for parent snapshot in snapshot pipeline nav, with a brief loading state and a set width to avoid making the tabs jump

Test Plan: View snapshot pages for different pipelines, verify that the "current"/"snapshot" tag shows up correctly.

Reviewers: alangenfeld, max, sandyryza

Differential Revision: https://dagster.phacility.com/D4789