RFC: Add partitions to arbitrary cron-based schedules
Summary:
With the magic of cronitor, there is no reason we can't generate a partition set for arbitrary cron schedules. This would allow us to maintain the invariant that all schedules have partition sets and can be backfilled. Just putting out for any initial thoughts/feedback.

The logic is backwards-compatible with existing XXXly_schedule partition logic: at a given execution time, the partition for the generated run is the previous execution time. This might not make sense for all schedules, so we could make it configurable if we want.

Test Plan: TBD

Reviewers: sashank, prha, alangenfeld, schrockn