Fix helm namespace lifecycle in tests
Summary:
I think the intention of this contextmanager is to leave the helm resources for the whole lifecycle, but currently its returning but not yielding so that doesn't happen. Make it a contextmanager so that it will wait until the end of the test to cleanup.

Test plan:
k8s integration tests now pass locally again

Reviewers: catherinewu, nate, johann