(config-scaffold 2/n) Add scaffold config button to Dagit playground
Summary: This diff adds a "scaffold missing config" button to Dagit, which is inspired by the scaffold CLI command. It does the scaffolding client side, based on the config schema.

There are a few choices here that are debatable. I've highlighted them in the comments. Open to feedback about how to handle them.

Test Plan: Created solid and resource with every possible config type. Checked to make sure scaffolded config was correct. Video:

Reviewers: schrockn, bengotow, dish

Subscribers: