Append "_location" to default repository location name
Summary:
This is to support @dish's refactor that is going to include repo location names in more places, and having the location name identical to the repo name is confusing. Could also shorten to '_loc' or something too if this feels too verbose.

We don't use the repo location name anywhere persistent (e.g. repo origins or pipeline origins), so this should be just a cosmetic change.

Test Plan: BK (Hope this breaks something, if it doesnn't I will add a test)

Reviewers: dish, alangenfeld, sashank, prha

Subscribers: dish

Differential Revision: https://dagster.phacility.com/D4827