RFC: Allow pipeline to be used as composite_solid
Summary: This is an RFC of an approach to this problem.

The general plan would be to do someting like this diff, which
makes pipeline usable in contexts where composite solids are currently
used. This was surprisingly tractable. After this diff we could
codemod all the composite_solids to pipelines, which is cool. However
there would still be work to do make dagit work well, etc.

Right now this adds quite a bit of code dup to get things working. My
plan was to introduce a GraphDefinition base class and move
implementations up into the base class until CompositeSolidDefinition
are and PipelineDefinition are effectively merged (they would still
exist to account for different ctor order and backwards compat)

We could also pre-emptively prevent this code dup by introducing
GraphDefinition all in one fell swoop. However that will expand the
scope of this quite a bit.

Next steps would be to make the config system work with top-level
config, inputs, and outputs keys. Right now if you tried to
use execute_pipeline on a pipeline with a config mapping or an
unfulfilled input it would break.

Test Plan: BK

Differential Revision: https://dagster.phacility.com/D4847