Move bad names from warnings to hard errors
Summary: We started emitting warnings for malformed names a few weeks
back. Given that we are doing other breaking changes this week
I am proposing pushing this one out.

The one thing I am retreating on is "nameless" pipelines. Instead
of hard failure we are instead going to generate a pipeline
with a name that conforms to the check. It is not a stable name.

Reviewers: alangenfeld, max

Test Plan: BK

Differential Revision: https://dagster.phacility.com/D4846