[celery-k8s] recover on celery task restart
Summary: From an observed log - if the celery workers are for example scaled down mid-task, the celery task is retried and we hit this error of the k8s job we intend to produce already existing.

Instead of exiting, if we just proceed we should gracefully recover.

The risk here is that the job name isn't the one we expect, but im not sure what conditions that could occur.

Test Plan: ?

Reviewers: #platform

Subscribers: