Change partition matrix click behavior, split large files and GraphQL types
Summary:
This diff expands on the partition matrix's "click to show runs" behavior so that clicking a particular step in a particulary day shows you all the runs of that day annotated with the state of that particular step and a link to go directly to the logs with the step filter applied.

I also spent some time cleaning up the Partition view React components to make it easier to do more performance related refactoring soon:

- I moved some supporting custom components to their own files in src/partitions. I'm a fan of keeping small custom components in the same file until they're used elsewhere, but PartitionView was getting out of hand.

- I gave the partition matrix and the graphs on the page distinct GraphQL fragments and rolled them up into the root query so that it's easier to trace where fields we're fetching are used. This revealed a few fields we weren't using and a few that are only needed for the pipeline-wide expectation results + materializations graphs, which we might cut soon.

In the spirit of trying to load as little data as possible, the UI no longer loads the data required for the RunTable until you click on a particular partition in the matrix. To render the table with the additional column showing the stats of the clicked step, it blends in the data from the original query (which fetched all the per-step stats, but less info about the runs.)

Test Plan: Kick the tires! Will write storybooks for PartitionMatrix this week

Reviewers: dish, prha, schrockn, sandyryza

Differential Revision: https://dagster.phacility.com/D4898