Mark pipeline as failed when there's a failure after launch but before execution starts
Summary: king reported this in https://github.com/dagster-io/dagster/issues/3016 - you'll get a hanging run if the gRPC server can't load the instance for some reason. Deal with that by catching failures in the instance.launch_run method and marking the run as failed when that happens. (Failures that happen during execution are responsible for marking the run as failed themselves - this is solely for the case where we weren't able to even start execution. We mark as failure in the run launching process because if it's an issue loading the instance, we have no way to find the run to mark it as failed.

Test Plan: New BK test

Reviewers: alangenfeld, johann, schrockn

Differential Revision: https://dagster.phacility.com/D4900