(make-pipeline-a-solid-4) Rename ISolidDefinition to NodeDefinition and spread name through composition process for code clarity
Summary: With the ability to embed pipelines/graphs within composition,
the use of the term "solid" through the process becomes more
problematic.

This renames ISolidDefinition to NodeDefinition to indicate something
that can embedded in a graph (which can be a graph or a solid) and
begins to rename the internal implementations to be "nodes" instead of
"solids". Otherwise variables with the name "solid" could actually
refer to pipelines/graph.

Depends on D4888

Reviewers: alangenfeld, max, sandyryza

Test Plan: BK

Differential Revision: https://dagster.phacility.com/D4912