Remove RepositoryLocationHandle static constructors, use RepositoryLocationOrigin instead
Summary: With the RepositoryLocationOrigin class introduced in https://dagster.phacility.com/D4916 we can make the handle class constructors just all take in an origin, do the creation in the constructor, and eliminate these unneeded static factory methods. Nothe main goal of creating the class (which was to make it possible to reload repositories that failed to load), but cleans things up a bit too.

Test Plan: BK

Reviewers: alangenfeld, sashank, schrockn, prha

Differential Revision: https://dagster.phacility.com/D4918