kill the cli api
Summary: A number of changes related to repo/pipeline/schedule origins become a lot simpler if the cli api goes away. This diff makes that happen - we can decide when we are comfortable doing this. Since 0.9.16 all users are no longer using the cli api by default, so this should be an invisible change to end users.

Test Plan: BK + Azure

Reviewers: alangenfeld, sashank, schrockn

Differential Revision: https://dagster.phacility.com/D4935