Summary: This reverts the creation of the dagster-aws-pyspark package and instead makes it possible to import dagster_aws.emr without transitively importing pyspark stuff.
Summary:
It accomplishes this by making parse_spark_configs generate a configs_spark.py in both the
dagster-aws and dagster-spark packages.

While this is probably not how we'd do it if we were starting over, it avoids a breaking change.

Depends on: D4938

Test Plan: bk + regenerated configs_spark.pys

Reviewers: schrockn

Differential Revision: https://dagster.phacility.com/D4939