RFC: add step_selection arg to execute_pipeline
Summary:
A difficulty that I am encountering when working on the slack digest pipeline: running a single step, including loading inputs

In this case, the step reads from the "events" table (because its parent OutputDefinition refers to an AssetStore that writes there) and writes to the "words" table.  I'd like to just execute it on whatever is in the message_events table.  The run that populated it is long gone, but I know that there's data there worth testing on.

If I try to execute it as a solid subset, this does not happen: the subset doesn't include the parent OutputDefinition, so I get a complaint that the input is missing and needs to be specified via config.

The workaround that we gave to a user for this issue was to use default_value. however, that doesn't really work in the AssetStore world - in pre-AssetStore world, the value for the input is just the name of the table.  In AssetStore world, the value for the input is the contents of the table.

Test Plan: need to add tests

Reviewers: alangenfeld, schrockn, yuhan

Subscribers: