Log original exception before trying to clean up a test instace
Summary: Right now if a test fails and then the cleanup fails it only logs the cleanup fail, making it hard to track down errors. Add the original exception to the logs before attempting to clean up.

Test Plan: Simulate a failure in poll_for_step_start(), run a test that calls it, see a much clearer error message

Reviewers: alangenfeld

Differential Revision: https://dagster.phacility.com/D4966