Show the original DB exception that triggerd a DagsterInstanceMigrationRequired on py2
Summary: We are being plagued by flaky DagsterInstanceMigrationRequired errors on various tests in py2 when we do innocuous things like read from the event log, and we have no way of seeing what the original DB exception was that triggered them. This diff ensures that we log the original exception info as well when this error fires - rather than using six.raise_from, which loses context on py2, pass in the original exception info and always log it.

Test Plan: BK

Reviewers: schrockn, max, alangenfeld, prha

Differential Revision: https://dagster.phacility.com/D4975