[dagit] Workspace paths
Summary:
Following the earlier proposal in the Quip doc.

For all objects that exist in the loaded workspace (e.g. anything in a repository/location), convert URL paths to use `/workspace` namespacing.

This way, all objects loaded in the current workspace are accessible via URL, rather than requiring the user to switch repository context via the repo switcher. When a repo@location path is loaded, the entire Dagit context changes to that repo, including the left nav. When switching repos via the left nav, the user lands on a view of all objects for that repo.

Path remappings:

- `/pipeline/...` -> `/workspace/repo@location/pipelines/...`
- `/solid/...` -> `/workspace/repo@location/solid/...`
- `/solids/...` -> `/workspace/repo@location/solids/...` (we should consolidate these solids routes...)
- `/schedules/...` -> `/workspace/repo@location/schedules/...`

This also necessitates the creation of disambiguation routes to allow the user to choose between objects with name collisions across repos. If only one object matches the name, we make our best guess and redirect to that object. For instance, `/workspace/pipelines/[pipeline_name]`.

Other notes:

- Instance routes (runs, assets, scheduler) are already in place.
- The `repoAddress` prop is passed through the workspace route components, guaranteeing that the `repo@location` value is available to all constructed URLs within those components.
  - This is so that we don't have to put null checking in place for URL construction, e.g. in cases where we are within an instance route (`/instance/runs...`) and have no URL repo information to build paths with.

Test Plan: Load Dagit, click everywhere on the site that I possibly can. Verify proper navigation and rendering of routes.

Reviewers: bengotow, alangenfeld, dgibson, prha, sashank

Differential Revision: https://dagster.phacility.com/D5084