Correctly test fall DST transition
Summary:
Realized (after living through it in 2010) that I had the times wrong here - in the spring, 2:30 AM doesn't exist, but it would be 1:30AM that gets repeated in the fall, not 2:30AM. So to test this correctly we need to change the time. Luckily it still seems to work fine (it picks the first 1:30 AM, not the second 1:30AM, which seems fine).

Test Plan: BK