clean up some places where we weren't using the fact that grpc location handles are contextmanagers
Summary:
Only one of these was a bug (where we were leaving a heartbeat thread hanging by not cleaning up the handle) - the rest were places where we could just use a with() statement but weren't for some reason.

Also just use a default instance in the grpc test now that gRPC is the default.

Test Plan: BK + Azure