Remove broken repo telemetry logic on dagit startup
summary:
This broke when we turned on gRPC by default, but it really hasn't worked correctly since 0.8.0 and containerization landed. rex is actively revamping it now, removing it in the meantime.

Test Plan:
Load dagit pointed at a remote gRPC server with a different working directory, no more error due to unloadable repo in this codepath

Reviewers: rexledesma, alangenfeld, prha, johann