[dagit] Correctly reload repository state
Summary:

We currently track the current repository in state and invalid that state only when the repo hash changes, which means that when we perform queries that would update a repo using the reload button, it doesn't pick up any new changes from the server because the hash hasn't changed.

Instead, track the repo hash in state, and use that to return the matching repo from our options list. The options list is always up to date from the query, so this should give us the correct repo state at all times.

Test Plan:

Load dagit, then remove a pipeline from the toys repo and click the reload button. Verify that the left nav changes correctly.

Reviewers:

dgibson, alangenfeld, sashank, prha

Subscribers: