RFC DagsterType.name -> DagsterType.unique_name
Summary: There are a bunch of bad error messages since `name` on `DagsterType` is this sort of confusing property related to uniqueness, and `display_name` is the thing you want.

To address this, `self.name` -> `self._name` and add a getter for `unique_name`, and update all the callsites to use either display or unique name accordingly.

This diff doesn't change the input arg to `DagsterType`, since i think that isn't as problematic and is more costly to change.

Test Plan: bk

Reviewers: schrockn, dgibson, max, rexledesma

Subscribers: