RFC use custom yaml loader
Summary:
If someone has a yaml file with a datestring like `2020-01-01` , it gets parsed in to a `datetime` object which will fail when it hits our `serdes` libraries.

I think if we want to support `datetime` object creation, it sould be part of the config system's post processing, not the yaml step.

This diff changes our yaml utils to use a custom child of `SafeLoader` which drops this `datetime` resolution.

resolves #2654

Test Plan: added test, can add more

Reviewers: schrockn, max, sandyryza

Subscribers: