fix test issue when test is running between midnight UTC and midnight central time
Summary:
This is purely a test issue - we call today_at_midnight() to determine a suitable start date for the schedule, but we were calculating the date to use in the local timezone of the machine. BK machines run in UTC (I think?) so we were setting a date that was late enough that the schedule had no partitions.

This diff fixes that by calculating the start date for that schedule using US/Central, so there were always be at least one partition.

Test Plan: BK (in the dark of night only)

Reviewers: sashank, alangenfeld, max