fix fan-in input mappings in composites
Summary:
This takes an approach of fixing things without changing the input/output mapping APIs. It's not obvious to me how the output mappings would work in that version of the world, and breaking changes have their own inherent cost. This approach is a little awkward, but I think maybe the way to go.

depends on D5217

Test Plan: added test cases

Reviewers: schrockn

Differential Revision: https://dagster.phacility.com/D5268