Add a hook that simulates React.useState but syncs with querystring
Summary: Overall I think that we should be doing this (saving important state to the query string so it’s preserved on page reload) more often. This is my stab at making a hook that replicates the setState interface so that there’s a lower barrier to implementing this. I feel like usually I just go with local state on the page because parsing the querystring from location.search is a bit of a pain.

Test Plan: I just clicked through all the existing uses I replaced - there's more use of this in my stacked diff!

Reviewers: dish, prha, schrockn, sandyryza

Subscribers: