ButtonLink
Summary:

Create a new generic component: a button that looks like a link. This is a common pattern for things that are not actually links to anything, but have onclick behavior and should *look* like links.

Test Plan:

View storybook examples, verify correct rendering and hover/focus behavior.

Reviewers:

bengotow, sashank

Subscribers: