[dagit] Use /solids for individual solid views
Summary:

Use the full `/solids` page for all individual solids views. The "typeExplorer" links currently don't work (or seem to mean anything) in an individual solid view, and it seems like we may as well show everything when looking at individual solids anyway.

This fixes the `typeExplorer` links in individual solid views, and makes the virtualized list scroll directly to the selected solid.

Reviewers: bengotow, max

Differential Revision: https://dagster.phacility.com/D5201