[dagit] Fix initial localStorage active repo state
Summary:

Fix 1 of 2 for #3254.

Use `localStorage` correctly when determining initial repository state in Dagit. We're currently overlooking this and just defaulting to the first repository in the retrieved array, which ends up resetting the state when someone opens a new tab (e.g. by kicking off a run).

Fix this by just initializing correctly.

Test Plan:

Lint, jest, ts.

Load Dagit, switch repositories. Open a new tab, verify that the localStorage state is preserved and used correectly to start Dagit on the repo that I switched to.

Reviewers:

bengotow, alangenfeld, prha

Subscribers: