[dagit] Header cleanup for instance roots
Summary:

Clean up the headers on instance roots, which currently use `TopNav`. We don't need the added gray chrome for these, and the page-level headers can be clearer with larger text.

I am in favor of using similar headers for object pages as well, as we've done for schedules and solids. This moves us in that direction.

Test Plan:

View `/instance/runs`, `/instance/assets`, `/instance/scheduler`, `/instance/details`.

Reviewers:

bengotow

Subscribers: