Refactor configured infrastructure to eliminate "underscore" params on public definitions
Summary: As I've been exploring the notion of
graph/composite_solid/pipeline and the notion of consolidating a bunch
of concepts into the resource system, I keep on running into the
configured stuff, which has some pretty rough edges.

In particular I think the inclusion of these internal parameters on public APIS
to implement the configured pattern is quite problematic.

Instead of checking for the recursion by passing around an opaque
closure that contains back pointers to the parent definition it its
implementation, instead we just construct a lineage of configuration
schemas as a data structure, and pass that data structure in where
config schema used to be allowed. The config evaluation then traces this
data structure. This allows us to reuse the config_schema param for
this purpose.

Although this does not do this, I am looking forward to potentially
completely merging the ConfigMapping notion on Graph with configured.
They are both stacks of config expansion and should share more code.

Test Plan: BK

Differential Revision: https://dagster.phacility.com/D5271

Reviewers: sandyryza, sashank, alangenfeld