RFC: loader_key
Summary:
Depends on D5285.

This happens to use AssetStores under the covers for convenience of prototyping.  However, from the perspective of the user, the connection isn't important.

The connection will become more important when in the "using a loader on the output of an upstream solid" case.

Test Plan: bk

Differential Revision: https://dagster.phacility.com/D5307