Add ScopedResourceSet to thread down through the config parsing code paths
Summary: I'm prototyping this as well and it was a bit of work to get this threaded through all the right places. We don't need to commit this but I thought it was a useful thing to arc patch so I'm throwing it up here. Did a class because it makes param checking easier, we will eventually need to remap resources as we thread it down that stack, and we can do things like have wrapper methods for accessing loaders and asset stores and assert that we are getting the right type of resource (they will be subclasses)

Test Plan: BK

Reviewers: sandyryza

Differential Revision: https://dagster.phacility.com/D5317