Cli for daemon heartbeat check
Summary:
This can be used in k8s and wherever else a liveness check is required.

This deviates a bit from our other liveliness checks (and I think the standard) in that instead of actually looking at the process inside the container, it asks the DB. I think I prefer this over adding some port on which the daemon health can be checked. This keeps us with a single source of truth for daemon liveliness. If the db goes down then this container will be considered unhealthy, but the container would be crashing anyway. I might be overlooking some pitfall though, so I'd appreciate feedback here.

Test Plan:
Integration

manually checked return code with and without the daemon running

manually did a helm install and checked that probe was working

Reviewers: dgibson, alangenfeld, catherinewu, nate

Reviewed By: catherinewu

Differential Revision: https://dagster.phacility.com/D5323