[dagit] Confirmation for run deletion
Summary:

Relates to #3236.

Add a confirmation dialog when batch deleting runs and remove `canTerminate` as a criteria for deletability.

When runs are successfully deleted, show a single toast. When any fail, show toasts for each of the failures.

Test Plan:

Delete runs from the Runs page, verify that the confirmation dialog appears and behaves correctly. Verify that deletion shows the appropriate toasts.

Reviewers:

alangenfeld, catherinewu, dgibson

Subscribers: