Add "--with-daemon" flag to dagit for local development?
Summary:
I'm torn about whether this is a good idea, putting it out for discussion. To make it so that you don't have to mess around with multiple terminal windows when developing locally on things that involve schedules, sensors, or queued runs, add a command-line arg to dagit that spawns the dagster-daemon process in the background as well. However, it's kind of bizarre to do this since it has nothing to do with dagit. We'd never want a user to do this in production since the only option if the daemon crashes for some reason is to restart dagit.

Test Plan: Will add BK coverage if this seems worth doing

Reviewers: alangenfeld, prha, johann, schrockn, nate