Add DockerRunLauncher
Summary:
This adds a run launcher that we could include as part of a docker-compose based deployment using a container for dagit and a container for dagster-daemon. (Our previous non-k8s docker example relied on everything being in a single container, which doesn't work well with the addition of the daemon). It's pretty similar to the existing celery docker executor, just launching runs instead of executing steps.

That said if we want to add this, I should probably create a separate dagster-docker module instead with no celery dependencies?

Test Plan: BK

Reviewers: alangenfeld, nate, sashank, johann, schrockn

Subscribers: prha, schrockn

Differential Revision: https://dagster.phacility.com/D5348