Eliminate required_resource_keys from ExecutorDefinition
Summary: This appears to be unused and appears to have no effect on execution from code inspection.

This is part of an exploration I am doing have executors and
intermediate storages become resources (through inheritance or a shared
interface) in preparation for the true merge of the concepts. I want to
make sure we add appropriate deprecations in the 0.10.0 to prepare for
anything we want to do on this from in 0.11.0. However while poking
around I might do cleanups this this.

We could either eliminate this straightaway, or we could add
required_resource_keys to ResourceDefinition and then do a merge.

Building on yuhan's stack because I am making changes down the line that
build that

Depends on D5339

Reviewers: max, alangenfeld, cdecarolis, yuhan

Test Plan: BK

Differential Revision: https://dagster.phacility.com/D5354